postcodes/pyproject.toml
[project]
name = "ukproperty-test"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.2.1",
    "geopandas>=1.0.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.1",
    "notebook>=7.3.2",
    "polars>=1.24.0",
    "pyarrow>=19.0.1",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
